finalisation de la tache

+++ src/api/api.service.ts
@@ -1,8 +1,8 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import { Avis, stuff } from './models/Stuff';
 import { Model, ObjectId } from 'mongoose';
-import { avisDto, stuffDto } from './models/stuffDto';
+import { avisDto, StuffDto } from './models/stuffDto';
 import { User } from '../auth/models/User';
 
 @Injectable()
@@ -10,7 +10,7 @@ export class ApiService {
 
     constructor(
         @InjectModel(stuff.name) private readonly stuffModel: Model<stuff>,
-        @InjectModel(Avis.name) private readonly avisModel: Model<Avis>,
+        @InjectModel(Avis.name) private readonly avisModel: Model<Avis>
     ) {}
 
     // recuperation de tous les objects
@@ -28,21 +28,32 @@ export class ApiService {
     }
     
     // poster des object
-    async postThing(thing: stuffDto, user: User): Promise<stuff> {
+    async postThing(thing: StuffDto, id: User): Promise<stuff> {
         // delete thing._id;
         // delete thing.userId;
         const data = Object.assign(thing);
-        return await new this.stuffModel(data).save();
+        return await new this.stuffModel({
+            ...data,
+            userId: id
+        }).save();
     }
 
     // supprimer un object
-    async deleteThing(thingId: string): Promise<stuff> {
+    async deleteThing(thingId: string, id: string): Promise<stuff> {
+        const stuff = await this.getOneThing(thingId);
+        if (stuff.userId.toString() !== id) {
+            throw new ForbiddenException('Vous n\'avez pas le droit de supprimer cette object!');
+        }
         return await this.stuffModel.findByIdAndDelete(thingId);
     }
 
     // Modification d'un objet
-    modifyThing(id: string, thing: stuffDto): Promise<stuff> {
+    async modifyThing(id: string, thing: StuffDto, idUser: string): Promise<stuff> {
         // delete thing._id;
+        const stuff = await this.getOneThing(id);
+        if (stuff.userId.toString() !== id) {
+            throw new ForbiddenException('Vous n\'avez pas le droit de modifier cette object!');
+        }
         return this.stuffModel.findByIdAndUpdate(id, thing, {
             new: true,
             runValidators: true
